Move the tests for the disk initialization module
Move the tests for the disk initialization module to a new file.
